Add blink1-webcam-busylight.sh
blink1-webcam-busylight.sh automatically changes a Blink(1)'s color
based on a USB Webcam's camera or audio use, turning the Blink(1)
into a handy "Busy Light" to notify others that a user is on Skype,
Google Chat or other similar services.

It was written as a quick proof of concept for a larger project
that I have planned.