Updated the code to use StatusType as oppose to StatusTitle
Should be using the type of status, rather than the title of the status.
The title of the status is in the unlimited space and may have different
messages with the same type. (i.e. Busy vs "Do not disturb"). Pidging
currently has 5 common types of statuses found at the following url

https://developer.pidgin.im/wiki/DbusHowto